Thanks for choosing to learn from EazyBytes

+++ section9/springsecsection9/src/main/java/com/eazybytes/repository/AccountsRepository.java
@@ -1,13 +1,12 @@
 package com.eazybytes.repository;
 
+import com.eazybytes.model.Accounts;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import com.eazybytes.model.Accounts;
-
 @Repository
 public interface AccountsRepository extends CrudRepository<Accounts, Long> {
-	
-	Accounts findByCustomerId(int customerId);
+
+    Accounts findByCustomerId(long customerId);
 
 }
