+++ section9/springsecsection9/src/main/java/com/eazybytes/repository/CardsRepository.java
@@ -2,6 +2,7 @@ package com.eazybytes.repository;
 
 import java.util.List;
 
+import com.eazybytes.model.Loans;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +11,6 @@ import com.eazybytes.model.Cards;
 @Repository
 public interface CardsRepository extends CrudRepository<Cards, Long> {
 	
-	List<Cards> findByCustomerId(int customerId);
+	List<Cards> findByCustomerId(long customerId);
 
 }
