+++ section9/springsecsection9/src/main/java/com/eazybytes/repository/LoanRepository.java
@@ -10,6 +10,6 @@ import com.eazybytes.model.Loans;
 @Repository
 public interface LoanRepository extends CrudRepository<Loans, Long> {
 	
-	List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+	List<Loans> findByCustomerIdOrderByStartDtDesc(long customerId);
 
 }
